% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{count.date}
\alias{count.date}
\title{count date}
\usage{
count.date(date, date.min = NULL, date.max = Sys.time(),
  by = "1 month")
}
\arguments{
\item{date}{A vector of character strings for dates.}

\item{date.min}{A character string for date, indicating the starting date.}

\item{date.max}{A character string for date, indicating the ending date.}

\item{by}{A character string. A increment of date. "1 month", "1 day", etc. are available.}
}
\description{
Make grouped data for detection data.
}
