% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{count.date2}
\alias{count.date2}
\title{count date 2}
\usage{
count.date2(
  ddate,
  cdate,
  date.min = NULL,
  date.max = Sys.time(),
  by = "1 month"
)
}
\arguments{
\item{ddate}{A vector of character strings for detection dates.}

\item{cdate}{A vector of character strings for correction dates.}

\item{date.min}{A character string for date, indicating the starting date.}

\item{date.max}{A character string for date, indicating the ending date.}

\item{by}{A character string. A increment of date. "1 month", "1 day", etc. are available.}
}
\description{
Make grouped data for detection and correction.
}
\note{
The format of date is yyyy-mm-dd. NA means that the correction has not been done by the max date.
}
