% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_logit.R
\docType{class}
\name{dGLM}
\alias{dGLM}
\alias{dGLM}
\alias{dGLM.logit}
\alias{dGLM.probit}
\alias{dGLM.cloglog}
\title{Class for NHPP-based software reliability model with d-metrics}
\format{\code{\link{R6Class}} object.}
\usage{
dGLM

dGLM.logit

dGLM.probit

dGLM.cloglog
}
\value{
Object of \code{\link{R6Class}} with methods for NHPP-based software reliability model with d-metrics.
}
\description{
Class for NHPP-based software reliability model with d-metrics
}
\section{Fields}{

\describe{
\item{\code{name}}{A character string for the name of model.}

\item{\code{params}}{A numeric vector for the model parameters.}

\item{\code{df}}{An integer for the degrees of freedom of the model.}

\item{\code{data}}{Data to esimate parameters.}
}}

\section{Methods}{

\describe{
  \item{\code{print()}}{This method prints model parameters.}
  \item{\code{omega()}}{This method returns the number of total faults.}
  \item{\code{coefficients()}}{This method returns a vector for the coefficients.}
  \item{\code{mvf(t, data = NULL)}}{This method returns the mean value function at time t.
         The d-metrics is given from \code{data}. If \code{data} is NULL, the d-metrics for the estimation is used.}
  \item{\code{dmvf(t, data = NULL)}}{This method returns the intensity function at time t.
         The d-metrics is given from \code{data}. If \code{data} is NULL, the d-metrics for the estimation is used.}
  \item{\code{residual(t, data = NULL)}}{This method returns the expected residual number of faults at time t.
         The d-metrics is given from \code{data}. If \code{data} is NULL, the d-metrics for the estimation is used.}
  \item{\code{ffp(t, data = NULL)}}{This method returns the fault-free probability at time t.
         The d-metrics is given from \code{data}. If \code{data} is NULL, the d-metrics for the estimation is used.}
  \item{\code{init_params(data)}}{This method changes the model parameters based on a given data.
         This is used to set the initial value for the fitting algorithm.}
  \item{\code{set_params(params)}}{This method sets the model parameters.}
  \item{\code{set_data(data)}}{This method sets data.}
  \item{\code{em(params, data)}}{This method returns a list with an updated parameter vector (param),
         absolute difference of parameter vector (pdiff),
         log-likelihood function for a given parameter vector (llf),
         the number of total faults (total) via EM algorithm for a given data.}
  \item{\code{llf(data)}}{This method returns the log-likelihood function for a given data.}
}
}

\seealso{
\code{\link{fit.srm.logit}}
}
\keyword{datasets}
