% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmetrics-data.R
\docType{data}
\name{dmet}
\alias{dmet}
\alias{dmet.ds1}
\alias{dmet.ds2}
\alias{dmet.ds3}
\title{Software fault data}
\description{
Software fault data with d-metrics (dynamic metrics)
}
\details{
This data set are software fault data with dynamic metrics such as the number of test cases, and so on.
in real embedded software which is control programs for a printer.
\describe{
 \item{ds1}{The fault data for control programs for a printer.
    The number of total faults is 66, and the number of testing periods is 20.
    Metrics are;
    \describe{
      \item{day}{The day of testing period.}
      \item{fault}{The number of faults detected at the testing period.}
      \item{tc}{The number of test cases executed at the testing period.}
      \item{ctc}{The cumulative number of test cases executed by the testing period.}
      \item{cov}{The gain of C0 coverage at the testing period.}
      \item{ccov}{The C0 coverage by the testing period.}
    }}
 \item{ds2}{The fault data for control programs for a printer.
    The number of total faults is 58, and the number of testing periods is 33.
    Metrics are;
    \describe{
      \item{day}{The day of testing period.}
      \item{fault}{The number of faults detected at the testing period.}
      \item{tc}{The number of test cases executed at the testing period.}
      \item{ctc}{The cumulative number of test cases executed by the testing period.}
      \item{cov}{The gain of C0 coverage at the testing period.}
      \item{ccov}{The C0 coverage by the testing period.}
    }}
 \item{ds3}{The fault data for control programs for a printer.
    The number of total faults is 52, and the number of testing periods is 30.
    Metrics are;
    \describe{
      \item{day}{The day of testing period.}
      \item{fault}{The number of faults detected at the testing period.}
      \item{tc}{The number of test cases executed at the testing period.}
      \item{ctc}{The cumulative number of test cases executed by the testing period.}
      \item{cov}{The gain of C0 coverage at the testing period.}
      \item{ccov}{The C0 coverage by the testing period.}
    }}
}
}
\references{
H. Okamura, Y. Etani, and T. Dohi (2010) A multi-factor software reliability model
based on logistic regression, Proceedings of 21st International Symposium on Software Reliability
Engineering (ISSRE2010), 31-40.
}
