% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dmvf}
\alias{dmvf}
\title{Difference of mean value function}
\usage{
dmvf(t, srm)
}
\arguments{
\item{t}{A numeric vector for time}

\item{srm}{An object of a model}
}
\value{
A vector of mean value function
}
\description{
This function provides the difference of mean value function for a given model.
}
\examples{
data(dacs)
result <- fit.srm.nhpp(fault=tohma, srm.names = c("exp", "gamma"),
            selection = NULL)
dmvf(c(1,2,3), result$exp)
}
