% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/em_nhpp.R
\name{em_exp_emstep}
\alias{em_exp_emstep}
\alias{em_gamma_emstep}
\alias{em_llogis_emstep}
\alias{em_lnorm_emstep}
\alias{em_lxvmax_emstep}
\alias{em_lxvmin_emstep}
\alias{em_pareto_emstep}
\alias{em_tlogis_emstep}
\alias{em_tnorm_emstep}
\alias{em_txvmax_emstep}
\alias{em_txvmin_emstep}
\alias{em}
\title{EM step}
\usage{
em_exp_emstep(params, data)

em_gamma_emstep(params, data, divide = 15L, eps = 1e-10)

em_llogis_emstep(params, data)

em_lnorm_emstep(params, data)

em_lxvmax_emstep(params, data)

em_lxvmin_emstep(params, data)

em_pareto_emstep(params, data)

em_tlogis_emstep(params, data)

em_tnorm_emstep(params, data)

em_txvmax_emstep(params, data)

em_txvmin_emstep(params, data)
}
\arguments{
\item{params}{A numeric vector for model parameters}

\item{data}{A faultdata}

\item{divide}{An integer for the number of integration points.}

\item{eps}{A numeric value for tolerance error.}

\item{...}{Other parameters}
}
\value{
A list with an updated parameter vector (param),
absolute difference of parameter vector (pdiff),
log-likelihood function for a given parameter vector (llf),
the number of total faults (total).
}
\description{
Execute one EM step for software reliability models.
}
