% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srm_poireg.R
\name{extractAIC.srm.poireg.result}
\alias{extractAIC.srm.poireg.result}
\title{AIC for a fit result of fit.srm.logit (S3 method)}
\usage{
\method{extractAIC}{srm.poireg.result}(fit, scale, k = 2, ...)
}
\arguments{
\item{fit}{An object of srm.logit.result.}

\item{scale}{used in the definition of the AIC statistic for selecting the
models. The default value, 0, indicates the scale should be estimated.}

\item{k}{the multiple of the number of degrees of freedom used for the
penalty. Only k = 2 gives the genuine AIC: k = log(n) is
sometimes referred to as BIC or SBC.}

\item{...}{Other parameters}
}
\description{
Return AIC for the fitting result.
}
\details{
This function is used in \code{step}.
}
