% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_logit.R
\name{faultdata.dmet}
\alias{faultdata.dmet}
\title{Software fault data with d-metrics}
\usage{
faultdata.dmet(formula, data, offset = NULL)
}
\arguments{
\item{formula}{An object of class formula. A symbolic description of the
model to be fitted. The output variable should be the column for the number
of faults.}

\item{data}{A dataframe for d-metrics and the number of faults.}

\item{offset}{An integer. This can be used to specify an a priori known
component to be included in the linear predictor during fitting. This should
be NULL or a numeric vector of length equal to the number of cases.}
}
\value{
A list with the attribute class='Rsrat.faultdata';
\item{fault}{A vector for the number of detected faults.}
\item{metrics}{A matrix for metrics.}
\item{total}{An integer for the number of total faults.}
\item{nmetrics}{An integer for the number of types of metrics.}
\item{offset}{A vector for offset.}
\item{len}{An integer for the number of data records.}
}
\description{
Function faultdata.dmet() creates a list to store the fault data and dymanic
metrics (d-metrics)that are used to esiamte model parameters of SRM.
}
\examples{
data(dmet)
faultdata.dmet(fault~., dmet.ds1)
}
