% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{faultdata}
\alias{faultdata}
\alias{faultdata.time}
\title{Software fault data}
\usage{
faultdata(time, fault = rep.int(0L, length(time)), type = rep.int(0L,
  length(time)))

faultdata.time(time, te)
}
\arguments{
\item{time}{A numeric vector for time intervals.}

\item{fault}{An integer vector for the number of faults detected in time intervals.
The fault detected just at the end of time interal is not counted.}

\item{type}{Either 0 or 1. If 1, a fault is detected just at the end of corresponding time interval.
This is used to represent the fault time data. If 0, no fault is detected at the end of interval.}

\item{te}{A numeric value for the time interval from the last fault to the observation time.}
}
\value{
A list with the attribute class='Rsrat.faultdata';
\item{time}{A vector for time interval.}
\item{fault}{A vector for the number of detected faults in intervals.}
\item{type}{A vector for the indicator whether a fault is detected at the end of intervals.}
\item{total}{An integer for the number of total faults.}
\item{len}{An integer for the number of time intervals.}
\item{mean}{A numeric value for mean fault detection time from data.}
\item{max}{A numeric value for maximum of fault detection time.}
}
\description{
Function faultdata() creates a list to store the fault data that are used to
esiamte model parameters of SRM. This data can represent both fault time and
count data.
}
\section{Functions}{
\itemize{
\item \code{faultdata.time}: Create fault data from fault time data.
}}

