% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataform_poireg.R
\name{faultdata.smet}
\alias{faultdata.smet}
\title{Software fault data with s-metrics}
\usage{
faultdata.smet(formula, data, names = NULL, offset = NULL)
}
\arguments{
\item{formula}{An object of class formula. A symbolic description of the
model to be fitted. The output variable should be empty.}

\item{data}{A dataframe for s-metrics.}

\item{names}{A list for the names of components. This is prioritized to the others.}

\item{offset}{An integer. This can be used to specify an a priori known
component to be included in the linear predictor during fitting. This should
be NULL or a numeric vector of length equal to the number of cases.}
}
\value{
A list with the attribute class='Rsrat.faultdata.smet';
\item{m}{An integer for the number of components.}
\item{names}{A list for the names of components.}
\item{metrics}{A matrix for s-metrics.}
\item{nmetrics}{An integer for the number of types of metrics.}
\item{offset}{A vector for offset.}
}
\description{
Function faultdata.smet() creates a list to store the fault data and dymanic
metrics (d-metrics)that are used to esiamte model parameters of SRM.
}
\details{
Since the row.names of data and names of srm.data are used to match the compoments,
it had better they all are labelled. If they are not labelled, the pairs are guessed
with their order.
}
\examples{
data(tomcat5)
faultdata.smet(~., tomcat5.smet)
}
