% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srm_logit.R
\name{fit.srm.logit}
\alias{fit.srm.logit}
\title{Regression-based software reliability model with d-metrics}
\usage{
fit.srm.logit(formula, data, linkfun = "logit", offset = NULL,
  control = list(), ...)
}
\arguments{
\item{formula}{An object of class formula. A symbolic description of the
model to be fitted. The output variable should be the column for the number
of faults.}

\item{data}{A dataframe for d-metrics and the number of faults.}

\item{linkfun}{A character string indicating a linkfunction. See Details.}

\item{offset}{An integer. This can be used to specify an a priori known
component to be included in the linear predictor during fitting. This should
be NULL or a numeric vector of length equal to the number of cases.}

\item{control}{A list of control parameters. See Details.}

\item{...}{Other parameters.}
}
\value{
A list with components;
\item{initial}{A vector for initial parameters.}
\item{srm}{A class of NHPP. The SRM with the estiamted parameters.}
\item{llf}{A numeric value for the maximum log-likelihood function.}
\item{df}{An integer for degrees of freedom.}
\item{convergence}{A boolean meaning the alorigthm is converged or not.}
\item{iter}{An integer for the number of iterations.}
\item{aerror}{A numeric value for absolute error.}
\item{rerror}{A numeric value for relative error.}
\item{data}{The data used.}
\item{linkfun}{The linkfunction used.}
\item{formula}{The formula supplied.}
\item{ctime}{A numeric value for computation time.}
\item{terms}{The terms object used.}
\item{call}{The method call.}
}
\description{
Estimate model parameters for d-metrics-based software reliability model.
}
\details{
The control argument is a list that can supply any of the following components:
\describe{
  \item{maxiter}{An integer for the maximum number of iterations in the fitting algorithm.}
  \item{reltol}{A numeric value. The algorithm stops if the relative error is
less than \emph{reltol} and the absolute error is less than \emph{abstol}.}
  \item{abstol}{A numeric value. The algorithm stops if the relative error is
less than \emph{reltol} and the absolute error is less than \emph{abstol}.}
  \item{stopcond}{A character string. \emph{stopcond} gives the criterion
for the stop condition of the algorithm. Either llf or parameter is selected.}
  \item{printflag}{A logical. If TRUE, the intermediate parameters are printed.}
  \item{printsteps}{An integer for print.}
}
The linkfun argument can take the following strings:
\describe{
  \item{logit}{A logit function.}
  \item{probit}{A probit function.}
  \item{cloglog}{A complementary log-log function.}
}
}
