% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srm_nhpp.R
\name{fit.srm.nhpp}
\alias{fit.srm.nhpp}
\title{NHPP-based software reliability model}
\usage{
fit.srm.nhpp(time, fault, type, te, data = data.frame(),
  srm.names = srm.models, selection = "AIC", control = list(), ...)
}
\arguments{
\item{time}{A numeric vector for time intervals.}

\item{fault}{An integer vector for the number of faults detected in time intervals.
The fault detected just at the end of time interal is not counted.}

\item{type}{Either 0 or 1. If 1, a fault is detected just at the end of corresponding time interval.
This is used to represent the fault time data. If 0, no fault is detected at the end of interval.}

\item{te}{A numeric value for the time interval from the last fault to the observation time.}

\item{data}{A dataframe. The arguments; time, fault, type, te can also be selected as the columns of dataframe.}

\item{srm.names}{A character vector, indicating the model (\code{\link{srm.models}}).}

\item{selection}{A character string, indicating the model selection criterion. The default is "AIC".
If this is NULL, the method returns the results for all model candidates.}

\item{control}{A list of control parameters. See Details.}

\item{...}{Other parameters.}
}
\value{
A list with components;
\item{initial}{A vector for initial parameters.}
\item{srm}{A class of NHPP. The SRM with the estiamted parameters.}
\item{llf}{A numeric value for the maximum log-likelihood function.}
\item{df}{An integer for degrees of freedom.}
\item{convergence}{A boolean meaning the alorigthm is converged or not.}
\item{iter}{An integer for the number of iterations.}
\item{aerror}{A numeric value for absolute error.}
\item{rerror}{A numeric value for relative error.}
\item{ctime}{A numeric value for computation time.}
\item{call}{The method call.}
}
\description{
Estimate model parameters for NHPP-based software reliability models.
}
\details{
The control argument is a list that can supply any of the following components:
\describe{
  \item{maxiter}{An integer for the maximum number of iterations in the fitting algorithm.}
  \item{reltol}{A numeric value. The algorithm stops if the relative error is
less than \emph{reltol} and the absolute error is less than \emph{abstol}.}
  \item{abstol}{A numeric value. The algorithm stops if the relative error is
less than \emph{reltol} and the absolute error is less than \emph{abstol}.}
  \item{stopcond}{A character string. \emph{stopcond} gives the criterion
for the stop condition of the algorithm. Either llf or parameter is selected.}
  \item{printflag}{A logical. If TRUE, the intermediate parameters are printed.}
  \item{printsteps}{An integer for print.}
}
}
\examples{
data(dacs)
fit.srm.nhpp(time=sys1[sys1>=0], te=-sys1[sys1<0], srm.names = c("exp"))
fit.srm.nhpp(fault=tohma, srm.names = c("llogis"))
}
