% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srm_poireg.R
\name{fit.srm.poireg}
\alias{fit.srm.poireg}
\alias{srm.poireg.options}
\title{Regression-based software reliability model with s-metrics}
\usage{
fit.srm.poireg(formula, data, srms, names = NULL, linkfun = "log",
  offset = NULL, control = list(), ...)

srm.poireg.options()
}
\arguments{
\item{formula}{An object of class formula. A symbolic description of the
model to be fitted. The output variable should be the column for the number
of faults.}

\item{data}{A dataframe for s-metrics.}

\item{srms}{A list of SRMs. The classes of NHPP, dGLM and results of fit.srm.nhpp and fit.srm.logit are allowed.}

\item{names}{A list of character strings to indicate the project names.}

\item{linkfun}{A character string indicating a linkfunction. See Details.}

\item{offset}{An integer. This can be used to specify an a priori known
component to be included in the linear predictor during fitting. This should
be NULL or a numeric vector of length equal to the number of cases.}

\item{control}{A list of control parameters. See Details.}

\item{...}{Other parameters.}
}
\value{
A list with components;
\item{initial}{A vector for initial parameters.}
\item{srm}{A class of NHPP. The SRM with the estiamted parameters.}
\item{llf}{A numeric value for the maximum log-likelihood function.}
\item{df}{An integer for degrees of freedom.}
\item{convergence}{A boolean meaning the alorigthm is converged or not.}
\item{iter}{An integer for the number of iterations.}
\item{aerror}{A numeric value for absolute error.}
\item{rerror}{A numeric value for relative error.}
\item{data}{The data used.}
\item{linkfun}{The linkfunction used.}
\item{formula}{The formula supplied.}
\item{ctime}{A numeric value for computation time.}
\item{terms}{The terms object used.}
\item{call}{The method call.}

A list of options.
}
\description{
Estimate model parameters for s-metrics-based software reliability models.

Generate a list of option values.
}
\details{
The control argument is a list that can supply any of the following components:
\describe{
  \item{maxiter}{An integer for the maximum number of iterations in the fitting algorithm.}
  \item{reltol}{A numeric value. The algorithm stops if the relative error is
less than \emph{reltol} and the absolute error is less than \emph{abstol}.}
  \item{abstol}{A numeric value. The algorithm stops if the relative error is
less than \emph{reltol} and the absolute error is less than \emph{abstol}.}
  \item{stopcond}{A character string. \emph{stopcond} gives the criterion
for the stop condition of the algorithm. Either llf or parameter is selected.}
  \item{trace}{A logical. If TRUE, the intermediate parameters are printed.}
  \item{printsteps}{An integer for print.}
}
The linkfun argument can take the following strings:
\describe{
  \item{log}{A log function.}
  \item{identity}{An identity function.}
}
}
\examples{
data(tomcat5)
data.tomcat5 <- list(catalina=tomcat5.catalina,
                     connector=tomcat5.connector,
                     jasper=tomcat5.jasper,
                     servlets=tomcat5.servlets,
                     tester=tomcat5.tester,
                     webapps=tomcat5.webapps)
results.tomcat5 <- lapply(data.tomcat5,
 function(d) fit.srm.nhpp(time=d$time, fault=d$fault, srm.names=c("gamma")))
fit.srm.poireg(formula=~LOC+St, data=tomcat5.smet, srms=results.tomcat5)
}
