% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{get.issues.github}
\alias{get.issues.github}
\title{Issues in GitHub}
\usage{
get.issues.github(repository, query, user = NULL, password = "",
  authtype = "basic", url = "https://api.github.com/repos",
  startPage = 1L, perPage = 30L)
}
\arguments{
\item{repository}{A character string indicating a repository such as "hadley/devtools"}

\item{query}{A character string for a query}

\item{user}{A string for user}

\item{password}{A string for password}

\item{authtype}{A string for authentication. The default is \code{basic}.}

\item{url}{An URL for GitHub}

\item{startPage}{An integer to identify the start page}

\item{perPage}{An integer to the number of issues for a page}
}
\description{
Collect issues from GitHub with a query
}
\examples{
## get crated date of blocer, critical and major issues
## which affected to 1.0.0 of ACE from Apache JIRA
## query <- "project = ACE AND issuetype = Bug"
## query <- paste(query, "AND priority in (Blocker, Critical, Major)")
## query <- paste(query, "AND affectedVersion = 1.0.0")
## get.issues.jira(query = query, fields="created")
}
