% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{get.issues.jira}
\alias{get.issues.jira}
\title{Issues in JIRA}
\usage{
get.issues.jira(query, userpwd,
  url = "https://issues.apache.org/jira/rest/api/2/search", startAt = 0,
  maxResults = 50)
}
\arguments{
\item{query}{A character string for a query}

\item{userpwd}{A string for user:passwd}

\item{url}{An URL for JIRA}

\item{startAt}{An integer to start}

\item{maxResults}{An integer to the number of issues for a transaction.}
}
\description{
Collect issues from JIRA with a query
}
\examples{
## get issuetype, priority, created, resolutiondate, resolution and versions of issues of ACE from Apache JIRA
get.issues.jira("project=ACE&fields=id,key,issuetype,priority,created,resolutiondate,resolution,versions",
    userpwd = "username:password")
}
