% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping-github.R
\name{get.opendate.github}
\alias{get.opendate.github}
\title{get open date from GitHub}
\usage{
get.opendate.github(repo, query = "state=all&labels=bug", user = NULL,
  password = "", authtype = "basic",
  url = "https://api.github.com/repos", startPage = 1L,
  perPage = 100L, date.min = NULL, date.max = Sys.time(),
  by = "1 month")
}
\arguments{
\item{repo}{A character string indicating a repository such as "hadley/devtools"}

\item{query}{A character string for a query}

\item{user}{A string for user}

\item{password}{A string for password}

\item{authtype}{A string for authentication. The default is \code{basic}.}

\item{url}{An URL for GitHub}

\item{startPage}{An integer to identify the start page}

\item{perPage}{An integer to the number of issues for a page}

\item{date.min}{A character string for date, indicating the starting date.}

\item{date.max}{A character string for date, indicating the ending date.}

\item{by}{A character string. A increment of date. "1 month", "1 day", etc. are available.}
}
\description{
Collect open date of issues from GitHub with query
}
