% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{get.opendate.jira}
\alias{get.opendate.jira}
\title{get open date from JIRA}
\usage{
get.opendate.jira(query, userpwd,
  url = "https://issues.apache.org/jira/rest/api/2/search", startAt = 0,
  maxResults = 50, date.min = NULL, date.max = Sys.time(),
  by = "1 month")
}
\arguments{
\item{query}{A character string for a query}

\item{userpwd}{A string for user:passwd}

\item{url}{An URL for JIRA}

\item{startAt}{An integer to start}

\item{maxResults}{An integer to the number of issues for a transaction.}

\item{date.min}{A character string for date, indicating the starting date.}

\item{date.max}{A character string for date, indicating the ending date.}

\item{by}{A character string. A increment of date. "1 month", "1 day", etc. are available.}
}
\description{
Collect open date of issues from JIRA with query
}
