% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumbel.R
\name{gumbel.min}
\alias{gumbel.min}
\alias{dgumbel.min}
\alias{pgumbel.min}
\alias{qgumbel.min}
\title{Gumbel distribution (minimum)}
\usage{
dgumbel.min(x, loc = 0, scale = 1, log = FALSE)

pgumbel.min(q, loc = 0, scale = 1, lower.tail = TRUE, log.p = FALSE)

qgumbel.min(p, loc = 0, scale = 1, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{x, q}{A numeric vector of quantiles.}

\item{loc}{A numeric value of location parameter.}

\item{scale}{A numeric value of scale parameter.}

\item{log, log.p}{A logical; if TRUE, the probability p is given as log(p).}

\item{lower.tail}{A logical; if TRUE, the probability is P[X <= x], otherwise, P[X > x].}

\item{p}{A numeric vector of probabilities.}
}
\value{
'dgumble.min' gives the desity, 'pgumbel.min' gives the distribution, and
'qgumbel.min' gives the quantile function.
}
\description{
Density, distribution function, and quantile function for
Gumbel distribution (minimum).
}
\details{
The Gumbel distribution (minimum) has the cumulative distribution function
F(q) = 1-exp(-exp(-z)) where z = (-q-loc)/scale
}
