% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumbel.R
\name{gumbel}
\alias{gumbel}
\alias{dgumbel}
\alias{pgumbel}
\alias{qgumbel}
\title{Gumbel distribution (maximum)}
\usage{
dgumbel(x, loc = 0, scale = 1, log = FALSE)

pgumbel(q, loc = 0, scale = 1, lower.tail = TRUE, log.p = FALSE)

qgumbel(p, loc = 0, scale = 1, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{x, q}{A numeric vector of quantiles.}

\item{loc}{A numeric value of location parameter.}

\item{scale}{A numeric value of scale parameter.}

\item{log, log.p}{A logical; if TRUE, the probability p is given as log(p).}

\item{lower.tail}{A logical; if TRUE, the probability is P[X <= x], otherwise, P[X > x].}

\item{p}{A numeric vector of probabilities.}
}
\value{
'dgumble' gives the desity, 'pgumbel' gives the distribution, and
'qgumbel' gives the quantile function.
}
\description{
Density, distribution function, and quantile function for
Gumbel distribution (maximum).
}
\details{
The Gumbel distribution (maximum) has the cumulative distribution function
F(q) = exp(-exp(-z)) where z = (q-loc)/scale
}
