% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgumbel.R
\name{lgumbel.min}
\alias{lgumbel.min}
\alias{dlgumbel.min}
\alias{plgumbel.min}
\alias{qlgumbel.min}
\alias{rlgumbel.min}
\title{log-Gumbel distribution (Weibull)}
\usage{
dlgumbel.min(x, loclog = 0, scalelog = 1, log = FALSE)

plgumbel.min(q, loclog = 0, scalelog = 1, lower.tail = TRUE, log.p = FALSE)

qlgumbel.min(p, loclog = 0, scalelog = 1, lower.tail = TRUE, log.p = FALSE)

rlgumbel.min(n, loclog = 0, scalelog = 1)
}
\arguments{
\item{x, q}{A numeric vector of quantiles.}

\item{loclog}{A numeric value of location parameter for Gumbel distribution.}

\item{scalelog}{A numeric value of scale parameter for Gumbel distribution.}

\item{log, log.p}{A logical; if TRUE, the probability p is given as log(p).}

\item{lower.tail}{A logical; if TRUE, the probability is P[X <= x], otherwise, P[X > x].}

\item{p}{A numeric vector of probabilities.}
}
\value{
'dlgumble.min' gives the desity, 'plgumbel.min' gives the distribution, and
'qlgumbel.min' gives the quantile function.
}
\description{
Density, distribution function, and quantile function for
log-Gumbel (Weibull) distribution.
}
\details{
The log-Gumbel distribution (minimum) has the cumulative distribution function
F(q) = 1-exp(-exp(-z)) where z = (-log(q)-loc)/scale
}
