% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llogis.R
\name{llogis}
\alias{llogis}
\alias{dllogis}
\alias{pllogis}
\alias{qllogis}
\alias{rllogis}
\title{log-logistic distribution}
\usage{
dllogis(x, locationlog = 0, scalelog = 1, log = FALSE)

pllogis(q, locationlog = 0, scalelog = 1, lower.tail = TRUE, log.p = FALSE)

qllogis(p, locationlog = 0, scalelog = 1, lower.tail = TRUE, log.p = FALSE)

rllogis(n, locationlog = 0, scalelog = 1)
}
\arguments{
\item{x, q}{A numeric vector of quantiles.}

\item{locationlog}{A numeric value of location parameter for the logistic distribution.}

\item{scalelog}{A numeric value of scale parameter for the logistic distribution.}

\item{log, log.p}{A logical; if TRUE, the probability p is given as log(p).}

\item{lower.tail}{A logical; if TRUE, the probability is P[X <= x], otherwise, P[X > x].}

\item{p}{A numeric vector of probabilities.}
}
\value{
'dllogis' gives the desity, 'pllogis' gives the distribution, and
'qllogis' gives the quantile function.
}
\description{
Density, distribution function, and quantile function for
log-logistic distribution.
}
\details{
The log-logistic distribution has the cumulative distribution function
F(q) = G(log(q)) where G(q) is the cumulative distribution function
of the logistic distribution with location and scale.
}
