% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_musa.R
\docType{data}
\name{musa}
\alias{musa}
\alias{musa.sys1.time}
\alias{musa.sys1.group}
\alias{musa.sys17.time}
\alias{musa.sys17.group}
\title{Software fault data}
\description{
Software fault data (time data, grouped data) collected by John Musa
}
\details{
This data set are software fault data in real application, which was
collected by John Musa.
\describe{
 \item{sys1}{136 faults have been detected during system test and operations in the program with 21,700 instructions.}
 \item{sys17}{38 faults have been detected during system test in the program with 61,900 instructions.}
}
}
