% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mvf}
\alias{mvf}
\title{Mean value function}
\usage{
mvf(t, srm)
}
\arguments{
\item{t}{A numeric vector for time}

\item{srm}{An object of a model}
}
\value{
A vector of mean value function
}
\description{
This function provides the mean value function for a given model.
}
\examples{
data(dacs)
result <- fit.srm.nhpp(fault=tohma, srm.names = c("exp", "gamma"),
            selection = NULL)
mvf(c(1,2,3), result$exp)
}
