% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mvfplot}
\alias{mvfplot}
\title{Draw a graph of mean value functions}
\usage{
mvfplot(time, fault, type, te, mvf, xlab = "time", ylab = "# of faults",
  datalab = "data", xmax = NA, ymax = NA, colors = mmcolors, ...)
}
\arguments{
\item{time}{A numeric vector for time intervals.}

\item{fault}{An integer vector for the number of faults detected in time intervals.
The fault detected just at the end of time interal is not counted.}

\item{type}{Either 0 or 1. If 1, a fault is detected just at the end of corresponding time interval.
This is used to represent the fault time data. If 0, no fault is detected at the end of interval.}

\item{te}{A numeric value for the time interval from the last fault to the observation time.}

\item{mvf}{A list of NHPP classes. See details.}

\item{xlab}{A character string, indicating the label of x-axis.}

\item{ylab}{A character string, indicating the label of y-axis.}

\item{datalab}{A character string, indicating the label of data in legend.}

\item{xmax}{A numeric value, indicating the maximum value of xlim. If NA, the maximum value is automatically determined.}

\item{ymax}{A numeric value, indicating the maximum value of ylim. If NA, the maximum value is automatically determined.}

\item{colors}{A vector of strings, indicating the colors in plot.}

\item{...}{Other parameters.}
}
\value{
An object of ggplot.
}
\description{
This function draws a graph of mean value function with data and NHPP classes.
}
\details{
The mvf argument should be a list of NHPP classes. In the function, \code{mvf}
and \code{name} attributes are used. Therefore, the list with \code{mvf} and
\code{name} can be used as an element of the list of mvf argument.
The return value is an object of ggplot2, and thus a line is added like
\code{mvfplot(...) + stat_function(fun=mvf)}.
}
\examples{
data(tomcat5)

result <- fit.srm.nhpp(time=tomcat5.catalina$time, fault=tomcat5.catalina$fault,
            srm.names = c("exp", "gamma"), selection = "all")

mvfplot(time=tomcat5.catalina$time, fault=tomcat5.catalina$fault,
     mvf=lapply(result, function(s) s$srm))
}
