% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_nhpp.R
\docType{class}
\name{NHPP}
\alias{NHPP}
\alias{NHPP}
\alias{ExpSRM}
\alias{GammaSRM}
\alias{ParetoSRM}
\alias{TNormSRM}
\alias{LNormSRM}
\alias{TLogisSRM}
\alias{LLogisSRM}
\alias{TXVMaxSRM}
\alias{LXVMaxSRM}
\alias{TXVMinSRM}
\alias{LXVMinSRM}
\title{Class for NHPP-based software reliability model}
\format{\code{\link{R6Class}} object.}
\usage{
NHPP

ExpSRM

GammaSRM

ParetoSRM

TNormSRM

LNormSRM

TLogisSRM

LLogisSRM

TXVMaxSRM

LXVMaxSRM

TXVMinSRM

LXVMinSRM
}
\value{
Object of \code{\link{R6Class}} with methods for NHPP-based software reliability model.
}
\description{
Class for NHPP-based software reliability model
}
\section{Fields}{

\describe{
\item{\code{name}}{A character string for the name of model.}

\item{\code{params}}{A numeric vector for the model parameters.}

\item{\code{df}}{An integer for the degrees of freedom of the model.}

\item{\code{data}}{Data to esimate parameters.}
}}

\section{Methods}{

\describe{
  \item{\code{print()}}{This method prints model parameters.}
  \item{\code{omega()}}{This method returns the number of total faults.}
  \item{\code{mvf(t)}}{This method returns the mean value function at time t.}
  \item{\code{inv_mvf(x)}}{This method returns the time at which the mean value function attains x.}
  \item{\code{intensity(t)}}{This method returns the intensity function at time t.}
  \item{\code{reliab(t, s)}}{This method returns the software reliability at time t from the orign s.}
  \item{\code{residual(t)}}{This method returns the expected residual number of faults at time t.}
  \item{\code{ffp(t)}}{This method returns the fault-free probability at time t.}
  \item{\code{imtbf(t)}}{This method returns the instantaneous MTBF at time t.}
  \item{\code{cmtbf(t)}}{This method returns the cumulative MTBF at time t.}
  \item{\code{median(s, p = 0.5)}}{This method returns the time at which the software reliability attains the proability p from the orign s.}
  \item{\code{init_params(data)}}{This method changes the model parameters based on a given data. This is used to set the initial value for the fitting algorithm.}
  \item{\code{set_params(params)}}{This method sets the model parameters.}
  \item{\code{em(params, data)}}{This method returns a list with an updated parameter vector (param),
         absolute difference of parameter vector (pdiff),
         log-likelihood function for a given parameter vector (llf),
         the number of total faults (total) via EM algorithm for a given data. \emph{divide} in GammaSRM is the number of integration points.}
  \item{\code{llf(data)}}{This method returns the log-likelihood function for a given data.}
}
}

\seealso{
\code{\link{srm}}
}
\keyword{datasets}
