% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srm_logit.R
\name{nobs.srm.logit.result}
\alias{nobs.srm.logit.result}
\title{Extract the number of observations (S3 method)}
\usage{
\method{nobs}{srm.logit.result}(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Other parameters}
}
\description{
Extract the number of observations from a model fit. This is principally
intended to be used in computing BIC (see AIC).
}
\details{
This function is used in \code{step}.
}
