% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto2.R
\name{pareto2}
\alias{pareto2}
\alias{dpareto2}
\alias{ppareto2}
\alias{qpareto2}
\title{Pareto distribution (type 2)}
\usage{
dpareto2(x, shape = 1, scale = 1, log = FALSE)

ppareto2(q, shape = 1, scale = 1, lower.tail = TRUE, log.p = FALSE)

qpareto2(p, shape = 1, scale = 1, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{x, q}{A numeric vector of quantiles.}

\item{shape}{A numeric value of shape parameter.}

\item{scale}{A numeric value of scale parameter.}

\item{log, log.p}{A logical; if TRUE, the probability p is given as log(p).}

\item{lower.tail}{A logical; if TRUE, the probability is P[X <= x], otherwise, P[X > x].}

\item{p}{A numeric vector of probabilities.}
}
\value{
'dpareto2' gives the desity, 'ppareto2' gives the distribution, and
'qpareto2' gives the quantile function.
}
\description{
Density, distribution function, and quantile function for
Pareto (type 2) distribution.
}
\details{
The Pareto type 2distribution has the cumulative distribution function
F(q) = 1 - (scale / (q + scale))^shape
}
