% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataform_logit.R
\name{print.Rsrat.faultdata.dmet}
\alias{print.Rsrat.faultdata.dmet}
\title{Printing software fault data with d-metrics}
\usage{
\method{print}{Rsrat.faultdata.dmet}(x, ..., digits = NULL,
  quote = FALSE, right = TRUE, row.names = TRUE)
}
\arguments{
\item{x}{An object of Rsrat.faultdata.}

\item{...}{Other parameters}

\item{digits}{The minimum number of significant digits.}

\item{quote}{A logical, indicating whether or not entries are printed with quotes.}

\item{right}{A logical, indicating whether or not strings are right-aligned.}

\item{row.names}{A logical or a character vector, indicating whether or not row names are printed.}
}
\description{
Print a data frame.
}
\details{
This function calls print.data.frame to forms the metrics data.
}
