% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srm_logit.R
\name{print.srm.logit.result}
\alias{print.srm.logit.result}
\title{Print a fit result of fit.srm.logit (S3 method)}
\usage{
\method{print}{srm.logit.result}(x, digits = max(3, getOption("digits") - 3),
  ...)
}
\arguments{
\item{x}{An object of srm.logit.result.}

\item{digits}{The minimum number of significant digits.}

\item{...}{Other parameters}
}
\description{
Print the fitting result.
}
\details{
This function uses the print for dGLM.
}
