% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srm_poireg.R
\name{print.srm.poireg.result}
\alias{print.srm.poireg.result}
\title{Print a fit result of fit.srm.poireg (S3 method)}
\usage{
\method{print}{srm.poireg.result}(x, digits = max(3, getOption("digits")
  - 3), ...)
}
\arguments{
\item{x}{An object of srm.poireg.result.}

\item{digits}{The minimum number of significant digits.}

\item{...}{Other parameters}
}
\description{
Print the fitting result.
}
\details{
This function uses the print for dGLM.
}
