% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rate}
\alias{rate}
\title{Rate function}
\usage{
rate(t, srm)
}
\arguments{
\item{t}{A numeric vector for time}

\item{srm}{An object of a model}
}
\value{
A vector of rate function
}
\description{
This function provides the rate function for a given model.
}
\examples{
data(dacs)
result <- fit.srm.nhpp(fault=tohma, srm.names = c("exp", "gamma"),
            selection = NULL)
rate(c(1,2,3), result$exp)
}
