% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{Rsrat}
\alias{Rsrat}
\title{Rsrat: A Software Reliability Assessment Tool with NHPP-Based Models}
\description{
This package provides estimation programs for software reliability growth
models based on NHPP.
}
