% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_poireg.R, R/model_ppoireg.R
\docType{class}
\name{sGLM}
\alias{sGLM}
\alias{sGLM}
\alias{sGLM.log}
\alias{sGLM.identity}
\alias{sGLM.penalized}
\alias{sGLM.penalized.log}
\alias{sGLM.penalized.identity}
\title{Class for software reliability model with s-metrics}
\format{\code{\link{R6Class}} object.}
\usage{
sGLM

sGLM.log

sGLM.identity

sGLM.penalized

sGLM.penalized.log

sGLM.penalized.identity
}
\value{
Object of \code{\link{R6Class}} with methods for software reliability model with s-metrics.
}
\description{
Class for software reliability model with s-metrics
}
\section{Fields}{

\describe{
\item{\code{srms}}{A list for srms.}

\item{\code{params}}{A numeric vector for the regression coefficients.}

\item{\code{df}}{An integer for the degrees of freedom of the model (total).}

\item{\code{data}}{Data to esimate parameters.}
}}

\section{Methods}{

\describe{
  \item{\code{print()}}{This method prints model parameters.}
  \item{\code{omega()}}{This method returns a vector of the number of total faults.}
  \item{\code{coefficients()}}{This method returns a vector of the coefficients for s-metrics.}
  \item{\code{init_params(data)}}{This method changes the model parameters based on a given data.
         This is used to set the initial value for the fitting algorithm.}
  \item{\code{set_params(params)}}{This method sets the model parameters.}
  \item{\code{set_data(data)}}{This method sets data.}
  \item{\code{em(params, data)}}{This method returns a list with an updated parameter vector (param),
         absolute difference of parameter vector (pdiff),
         log-likelihood function for a given parameter vector (llf),
         the number of total faults (total) via EM algorithm for a given data.}
  \item{\code{llf(data)}}{This method returns the log-likelihood function for a given data.}
}
}

\seealso{
\code{\link{fit.srm.poireg}}
}
\keyword{datasets}
