% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srm.R
\docType{data}
\name{srm.models}
\alias{srm.models}
\title{Model names}
\format{An object of class \code{character} of length 11.}
\usage{
srm.models
}
\value{
A list of all software reliability models (exp, gamma,
pareto, tnorm, lnorm, tlogis, llogis, txvmax, lxvmax, txvmin, lxvmin).
}
\description{
All software reliablity models.
}
\details{
Return a list of all models.
\describe{
\item{exp}{Exponential model (Goel-Okumoto model).}
\item{gamma}{Gamma model.}
\item{pareto}{Pareto model (Littlewood model).}
\item{tnorm}{Truncated normal model.}
\item{lnorm}{Log-normal model.}
\item{tlogis}{Truncated logisitc model (Inflection S-shpaed model, Logistic curve).}
\item{txvmax}{Truncated exttreme-value max model (Gompertz curve).}
\item{lxvmax}{Log-exttreme-value max model (Frechet typed).}
\item{txvmin}{Truncated exttreme-value min model.}
\item{lxvmin}{Log-exttreme-value min model (Goel model, Weibull model).}
}
}
\keyword{datasets}
