% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srm.R
\name{srm}
\alias{srm}
\title{Software reliability model}
\usage{
srm(names)
}
\arguments{
\item{names}{A character vector, indicating the model (\code{\link{srm.models}}).}
}
\value{
A class of NHPP (\code{\link{NHPP}}).
}
\description{
Generate a software reliability model.
}
\details{
Return a class of NHPP.
}
