% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgumbel.R
\name{tgumbel.min}
\alias{tgumbel.min}
\alias{dtgumbel.min}
\alias{ptgumbel.min}
\alias{qtgumbel.min}
\alias{rtgumbel.min}
\title{truncated Gumbel distribution (minimum)}
\usage{
dtgumbel.min(x, loc = 0, scale = 1, log = FALSE)

ptgumbel.min(q, loc = 0, scale = 1, lower.tail = TRUE, log.p = FALSE)

qtgumbel.min(p, loc = 0, scale = 1, lower.tail = TRUE, log.p = FALSE)

rtgumbel.min(n, loc = 0, scale = 1)
}
\arguments{
\item{x, q}{A numeric vector of quantiles.}

\item{loc}{A numeric value of location parameter for Gumbel distribution.}

\item{scale}{A numeric value of scale parameter for Gumbel distribution.}

\item{log, log.p}{A logical; if TRUE, the probability p is given as log(p).}

\item{lower.tail}{A logical; if TRUE, the probability is P[X <= x], otherwise, P[X > x].}

\item{p}{A numeric vector of probabilities.}
}
\value{
'dtgumble.min' gives the desity, 'ptgumbel.min' gives the distribution, and
'qtgumbel.min' gives the quantile function.
}
\description{
Density, distribution function, and quantile function for
truncated Gumbel distribution (minimum).
}
\details{
The truncated Gumbel distribution (maximum) has the cumulative distribution function
F(q) = 1-(1-G(q))/(1-G(0)), G(q) = exp(-exp(-z)) where z = (q-loc)/scale
}
