% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlogis.R
\name{tlogis}
\alias{tlogis}
\alias{dtlogis}
\alias{ptlogis}
\alias{qtlogis}
\title{truncated logistic distribution}
\usage{
dtlogis(x, location = 0, scale = 1, log = FALSE)

ptlogis(q, location = 0, scale = 1, lower.tail = TRUE,
  log.p = FALSE)

qtlogis(p, location = 0, scale = 1, lower.tail = TRUE,
  log.p = FALSE)
}
\arguments{
\item{x, q}{A numeric vector of quantiles.}

\item{location}{A numeric value of location parameter for the logistic distribution.}

\item{scale}{A numeric value of scale parameter for the logistic distribution.}

\item{log, log.p}{A logical; if TRUE, the probability p is given as log(p).}

\item{lower.tail}{A logical; if TRUE, the probability is P[X <= x], otherwise, P[X > x].}

\item{p}{A numeric vector of probabilities.}
}
\value{
'dtlogis' gives the desity, 'ptlogis' gives the distribution, and
'qtlogis' gives the quantile function.
}
\description{
Density, distribution function, and quantile function for
truncated logistic distribution.
}
\details{
The truncated logistic distribution has the cumulative distribution function
F(q) = 1-(1-G(q))/(1-G(0)) where G(q) is the cumulative distribution function
of the logistic distribution with location and scale.
}
