% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnorm.R
\name{tnorm}
\alias{tnorm}
\alias{dtnorm}
\alias{ptnorm}
\alias{qtnorm}
\title{truncated normal distribution}
\usage{
dtnorm(x, mean = 0, sd = 1, log = FALSE)

ptnorm(q, mean = 0, sd = 1, lower.tail = TRUE, log.p = FALSE)

qtnorm(p, mean = 0, sd = 1, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{x, q}{A numeric vector of quantiles.}

\item{mean}{A numeric value of mean parameter for the normal distirbution.}

\item{sd}{A numeric value of standard deviation for the normal distirbution.}

\item{log, log.p}{A logical; if TRUE, the probability p is given as log(p).}

\item{lower.tail}{A logical; if TRUE, the probability is P[X <= x], otherwise, P[X > x].}

\item{p}{A numeric vector of probabilities.}
}
\value{
'dtnorm' gives the desity, 'ptnorm' gives the distribution, and
'qtnorm' gives the quantile function.
}
\description{
Density, distribution function, and quantile function for
truncated normal distribution.
}
\details{
The truncated normal distribution has the cumulative distribution function
F(q) = 1-(1-G(q))/(1-G(0)), where G(q) is the cumulative distribution function of
normal distribution with mean and sd.
}
