% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dmvfplot}
\alias{dmvfplot}
\title{Draw a graph of diff of mean value functions on discrete time domain}
\usage{
dmvfplot(
  time = NULL,
  fault = NULL,
  type = NULL,
  te = NULL,
  data = data.frame(),
  srms = list(),
  xlab = "time",
  ylab = "# of faults",
  datalab = "data",
  xmax = NA,
  ymax = NA,
  colors = mmcolors,
  ...
)
}
\arguments{
\item{time}{A numeric vector for time intervals.}

\item{fault}{An integer vector for the number of faults detected in time intervals.
The fault detected just at the end of time interval is not counted.}

\item{type}{Either 0 or 1. If 1, a fault is detected just at the end of corresponding time interval.
This is used to represent the fault time data. If 0, no fault is detected at the end of interval.}

\item{te}{A numeric value for the time interval from the last fault to the observation time.}

\item{data}{A dataframe. The arguments; time, fault, type, te can also be selected as the columns of dataframe.}

\item{srms}{A list of estimated results.}

\item{xlab}{A character string, indicating the label of x-axis.}

\item{ylab}{A character string, indicating the label of y-axis.}

\item{datalab}{A character string, indicating the label of data in legend.}

\item{xmax}{A numeric value, indicating the maximum value of xlim. If NA, the maximum value is automatically determined.}

\item{ymax}{A numeric value, indicating the maximum value of ylim. If NA, the maximum value is automatically determined.}

\item{colors}{A vector of strings, indicating the colors in plot.}

\item{...}{Other parameters.}
}
\value{
An object of ggplot.
}
\description{
This function draws a graph of mean value function with data and NHPP classes.
}
\details{
The return value is an object of ggplot2, and thus a line is added like
\code{dmvfplot(...) + geom_point(stat="identity", position="identity", aes_string(x="x", y="exp", colour=shQuote("exp")))
 + geom_line(stat="identity", position="identity", aes_string(x="x", y="exp", colour=shQuote("exp")))}.
}
\examples{
data(dacs)
result <- fit.srm.nhpp(fault=tohma, srm.names = c("exp", "gamma"),
            selection = NULL)
dmvfplot(fault=tohma, srms=result)
}
