% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_nhpp.R
\name{em}
\alias{em}
\alias{em.exp}
\alias{em.gamma}
\alias{em.pareto}
\alias{em.tnorm}
\alias{em.lnorm}
\alias{em.tlogist}
\alias{em.llogist}
\alias{em.txvmax}
\alias{em.lxvmax}
\alias{em.txvmin}
\alias{em.lxvmin}
\title{EM step}
\usage{
em.exp(params, data, ...)

em.gamma(params, data, divide, ...)

em.pareto(params, data, ...)

em.tnorm(params, data, ...)

em.lnorm(params, data, ...)

em.tlogist(params, data, ...)

em.llogist(params, data, ...)

em.txvmax(params, data, ...)

em.lxvmax(params, data, ...)

em.txvmin(params, data, ...)

em.lxvmin(params, data, ...)
}
\arguments{
\item{params}{A numeric vector for model parameters}

\item{data}{A faultdata}

\item{...}{Other parameters}

\item{divide}{An integer for the number of integration points.}
}
\value{
A list with an updated parameter vector (param),
absolute difference of parameter vector (pdiff),
log-likelihood function for a given parameter vector (llf),
the number of total faults (total).
}
\description{
Execute one EM step for software reliability models.
}
