% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/em_nhpp.R
\name{em_exp_emstep}
\alias{em_exp_emstep}
\alias{em_gamma_emstep}
\alias{em_llogis_emstep}
\alias{em_llogis_estep}
\alias{em_llogis_pllf}
\alias{em_lnorm_emstep}
\alias{em_lxvmax_estep}
\alias{em_lxvmax_pllf}
\alias{em_lxvmin_estep}
\alias{em_lxvmin_pllf}
\alias{em_pareto_emstep}
\alias{em_tlogis_emstep_mo}
\alias{em_tlogis_estep}
\alias{em_tlogis_pllf}
\alias{em_tnorm_emstep}
\alias{em_txvmax_emstep_mo}
\alias{em_txvmax_estep}
\alias{em_txvmax_pllf}
\alias{em_txvmin_estep}
\alias{em_txvmin_pllf}
\alias{em}
\title{EM step}
\usage{
em_exp_emstep(params, data)

em_gamma_emstep(params, data, divide = 15L, eps = 1e-10)

em_llogis_emstep(params, data)

em_llogis_estep(params, data)

em_llogis_pllf(params, data, w1)

em_lnorm_emstep(params, data)

em_lxvmax_estep(params, data)

em_lxvmax_pllf(params, data, w1)

em_lxvmin_estep(params, data)

em_lxvmin_pllf(params, data, w1)

em_pareto_emstep(params, data)

em_tlogis_emstep_mo(params, data)

em_tlogis_estep(params, data)

em_tlogis_pllf(params, data, w0, w1)

em_tnorm_emstep(params, data)

em_txvmax_emstep_mo(params, data)

em_txvmax_estep(params, data)

em_txvmax_pllf(params, data, w0, w1)

em_txvmin_estep(params, data)

em_txvmin_pllf(params, data, w0, w1)
}
\arguments{
\item{params}{A numeric vector for model parameters}

\item{data}{A faultdata}

\item{divide}{An integer for the number of integration points.}

\item{eps}{A numeric value for tolerance error.}

\item{w1}{A numeric value indicating the expected number of faults after te.}

\item{w0}{A numeric value indicating the expected number of faults before 0.}

\item{...}{Other parameters}
}
\value{
A list with an updated parameter vector (param),
absolute difference of parameter vector (pdiff),
log-likelihood function for a given parameter vector (llf),
the number of total faults (total).
}
\description{
Execute one EM step for software reliability models.
}
\details{
\code{em_tlogis_emstep_mo} is an emstep based on Marshall-Olkin-type (maximum) with Exp

\code{em_tnorm_emstep} has been modified by using EM for truncated distribution. Concretely,
when the distribution is truncated at origin, the expected value is given by
\deqn{\text{E}[h(X)|D] = \sum_{i=1}^k h(x_i) + \frac{k}{\overline{F}(0)} \int_{-\infty}^0 h(x) f(x) dx.}

\code{em_txvmax_emstep_mo} is an emstep based on Marshall-Olkin-type (maximum) with Exp
}
