% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emfit.R
\name{emfit}
\alias{emfit}
\title{Estimate SRM parameters}
\usage{
emfit(srm, data, initialize = TRUE, maxiter = 2000, rtol = 1e-06,
  atol = 0.001, termination = "llf")
}
\arguments{
\item{srm}{An NHPP model}

\item{data}{A faultdata}

\item{initialize}{Either TRUE or FALSE. If TRUE, the model parameters are
initilized with a given data before executing the fitting algorithm.}

\item{maxiter}{An integer for the maximum number of iterations in the fitting algorithm.}

\item{rtol}{A numeric value. The algorithm stops if the relative error is
less than \emph{rtol} and the absolute error is less than \emph{atol}.}

\item{atol}{A numeric value. The algorithm stops if the relative error is
less than \emph{rtol} and the absolute error is less than \emph{atol}.}

\item{termination}{A character string. \emph{termination} gives the criterion
for the stop condition of the algorithm. Either llf or parameter is selected.}
}
\value{
A list with components;
\item{x}{a stationary vector.}
\item{iter}{the number of iterations in GS and CG methods.}
}
\description{
Estimate SRM parameters
}
