% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping-github.R
\name{get.issues.github}
\alias{get.issues.github}
\title{Issues in GitHub}
\usage{
get.issues.github(
  repo,
  query = "",
  user = NULL,
  password = "",
  authtype = "token",
  url = "https://api.github.com/repos",
  startPage = 1L,
  perPage = 100L
)
}
\arguments{
\item{repo}{A character string indicating a repository such as "hadley/devtools"}

\item{query}{A character string for a query}

\item{user}{A string for user}

\item{password}{A string for password or personal access token}

\item{authtype}{A string for authentication. The default is \code{token} (personal access toke).}

\item{url}{An URL for GitHub}

\item{startPage}{An integer to identify the start page}

\item{perPage}{An integer to the number of issues for a page}
}
\description{
Collect issues from GitHub with a query
}
