% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/gumbel.R
\name{dgumbel}
\alias{dgumbel}
\alias{pgumbel}
\alias{qgumbel}
\alias{rgumbel}
\alias{gumbel}
\title{Gumbel distribution}
\usage{
dgumbel(x, loc = 0, scale = 1, log = FALSE, min = FALSE)

pgumbel(q, loc = 0, scale = 1, lower = TRUE, log = FALSE, min = FALSE)

qgumbel(p, loc = 0, scale = 1, lower = TRUE, log = FALSE, min = FALSE)

rgumbel(n, loc = 0, scale = 1, min = FALSE)
}
\arguments{
\item{x, q}{A numeric vector of quantiles.}

\item{loc}{A numeric value of location parameter.}

\item{scale}{A numeric value of scale parameter.}

\item{log}{A logical; if TRUE, the probability p is given as log(p).}

\item{min}{A logical; if TRUE, the result of minimum Gumbel is provided.}

\item{lower}{A logical; if TRUE, the probability is P[X <= x], otherwise, P[X > x].}

\item{p}{A numeric vector of probabilities.}
}
\value{
'dgumble' gives the desity, 'pgumbel' gives the distribution, and
'qgumbel' gives the quantile function.
}
\description{
Density, distribution function, and quantile function for
Gumbel distribution.
}
\details{
The Gumbel distribution (maximum) has the cumulative distribution function
F(q) = exp(-exp(-z)) where z = (q-loc)/scale
The Gumbel distribution (minimum) has the cumulative distribution function
F(q) = 1-exp(-exp(-z)) where z = (-q-loc)/scale
}
