% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{Rsrat}
\alias{Rsrat}
\alias{Rsrat-package}
\title{Rsrat: A Software Reliability Assessment Tool}
\description{
This package provides estimation programs for software reliability growth
models, metrics-based software reliability growth models with logistic
regression for dynamic metrics and Poisson regression for static metrics.
}
